Shopping UI 작업  - 굿즈 리스트 추가  - 굿즈 상세 화면 작업 진행 중

+++ app/src/main/java/com/ezen/lolketing/view/main/shop/Shop2Activity.java
@@ -61,7 +61,7 @@ public class Shop2Activity extends AppCompatActivity
 
         String group = shopDTO.getGroup();
         String name = shopDTO.getName();
-        int price = shopDTO.getPrice();
+        Long price = shopDTO.getPrice();
 
         firestore.collection("ShopEvent").get()
                 .addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
